Moves IP Policies into their own plugin module
Refactors plugin.py, moving IP policies and associated tests into their
own modules for enchanced plugin.py readability.
