+++ quark/plugin.py
@@ -41,6 +41,7 @@ from quark.db import api as db_api
 from quark.db import models
 from quark import exceptions as quark_exceptions
 from quark import network_strategy
+from quark.plugin_modules import ip_policies
 from quark.plugin_modules import mac_address_ranges
 from quark.plugin_modules import security_groups
 from quark import plugin_views as v
@@ -942,59 +943,6 @@ class Plugin(neutron_plugin_base_v2.NeutronPluginBaseV2,
 
         return v._make_ip_dict(address)
 
-    def create_ip_policy(self, context, ip_policy):
-        LOG.info("create_ip_policy for tenant %s" % context.tenant_id)
-
-        ipp = ip_policy["ip_policy"]
-
-        if not ipp.get("exclude"):
-            raise exceptions.BadRequest(resource="ip_policy",
-                                        msg="Empty ip_policy.exclude regions")
-
-        ipp["exclude"] = netaddr.IPSet(ipp["exclude"])
-        network_id = ipp.get("network_id")
-        subnet_id = ipp.get("subnet_id")
-
-        model = None
-        if subnet_id:
-            model = db_api.subnet_find(context, id=subnet_id, scope=db_api.ONE)
-            if not model:
-                raise exceptions.SubnetNotFound(id=subnet_id)
-        elif network_id:
-            model = db_api.network_find(context, id=network_id,
-                                        scope=db_api.ONE)
-            if not model:
-                raise exceptions.NetworkNotFound(id=network_id)
-        else:
-            raise exceptions.BadRequest(
-                resource="ip_policy",
-                msg="network_id or subnet_id unspecified")
-
-        if model["ip_policy"]:
-            raise quark_exceptions.IPPolicyAlreadyExists(
-                id=model["ip_policy"]["id"], n_id=model["id"])
-        model["ip_policy"] = db_api.ip_policy_create(context, **ipp)
-        return v._make_ip_policy_dict(model["ip_policy"])
-
-    def get_ip_policy(self, context, id):
-        LOG.info("get_ip_policy %s for tenant %s" % (id, context.tenant_id))
-        ipp = db_api.ip_policy_find(context, id=id, scope=db_api.ONE)
-        if not ipp:
-            raise quark_exceptions.IPPolicyNotFound(id=id)
-        return v._make_ip_policy_dict(ipp)
-
-    def get_ip_policies(self, context, **filters):
-        LOG.info("get_ip_policies for tenant %s" % (context.tenant_id))
-        ipps = db_api.ip_policy_find(context, scope=db_api.ALL, **filters)
-        return [v._make_ip_policy_dict(ipp) for ipp in ipps]
-
-    def delete_ip_policy(self, context, id):
-        LOG.info("delete_ip_policy %s for tenant %s" % (id, context.tenant_id))
-        ipp = db_api.ip_policy_find(context, id=id, scope=db_api.ONE)
-        if not ipp:
-            raise quark_exceptions.IPPolicyNotFound(id=id)
-        db_api.ip_policy_delete(context, ipp)
-
     def get_mac_address_range(self, context, id, fields=None):
         return mac_address_ranges.get_mac_address_range(context, id, fields)
 
@@ -1043,3 +991,15 @@ class Plugin(neutron_plugin_base_v2.NeutronPluginBaseV2,
     def update_security_group(self, context, id, security_group):
         return security_groups.update_security_group(context, id,
                                                      security_group)
+
+    def create_ip_policy(self, context, ip_policy):
+        return ip_policies.create_ip_policy(context, ip_policy)
+
+    def get_ip_policy(self, context, id):
+        return ip_policies.get_ip_policy(context, id)
+
+    def get_ip_policies(self, context, **filters):
+        return ip_policies.get_ip_policies(context, **filters)
+
+    def delete_ip_policy(self, context, id):
+        return ip_policies.delete_ip_policy(context, id)
